Add a format class for reading video files
Opening videos via the open transform will result in all frames of a video
being loaded into memory, and stored as matrices of a single template.
Modify defaultFormat to try to read a file as a video if it can't be read as an
image (since video files can also have various file extensions).

Signed-off-by: caotto <ottochar@gmail.com>