Update threading model for stream (again)
Change the threading model used in stream slightly, generally as a given frame
is processed a single thread will attempt to continue working on it as it
progresses through the processing stages. Frames will be queued if a single
thread stage is already occupied when the thread gets there.

This update can be used with a large total number of frames being processed
without, while still maintaining relatively smooth playback by giving later
stages higher priority in the qt thread pool.