Add an option to draw on images without cloning them
Add an inPlace argument to draw, and add a couple transforms for drawing the
values of properties at some named point on an image.

The inPlace argument (default false) controls whether or not an image is cloned
before it is drawn on. Not cloning the image means the transforms doesn't
exactly obey the const modifier on the input template (since it alters the
values pointed at by that matrix), which could lead to unexpected results in
some scenarios.

On the other hand, the clone operation is wasteful in some cases, and also not
cloning can be useful in scenarios such as drawing a set of detections
represented by different templates on the same image, so as a compromise we
have a default off option to draw in place.