Add a few example algorithms using Stream for video processing
DisplayVideo just shows the frames of a video sequentially, PerFrameDetection
does face/eye detection on every frame and displays the results, AgeGenderDemo
uses AgeRegressor and GenderClassifier for age/gender detection, and displays
the results.

These algorithms expect a gallery with the name of one (or more) video sources
video sources can be the location of a video (including the address of an ip
camera), or the index of a webcam.