Clarify use of Common::seedRNG
Common::seedRNG uses a static variable to ensure that the RNG is seeded exactly
once per run of OpenBR. In spite of this, seedRNG was called from several
places, although the function would have an effect exactly once.

This commit
adds thread safety to seedRNG (previously the use of a static variable was
unsafe, although it could only cause a rather minor problem). More importantly
call seedRNG only from Context::initialize. If it is desirable to call seedRNG
more than once, changes will need to be made to support doing that in a
meaningful way.