Model separate project calls as seaprate list items in train
For non-leaf transforms, add a separate train method that takes
QList<TemplateList> as arguments, where the separate list items correspond to
the separate calls to project that would be made by parent transforms during
enrollment. This is basically to deal with the effects of distribute on
untrainable meta transforms such as flatten, we have to maintain a consistent
grouping with project, otherwise the intermediate projection results used
during training won't be consistent with what they would be during enrollment.