Try to recover if creating a QApplication fails
If we try to create a QApplication and fail, qt will call qFatal. We cannot
recover from that because even if we install a message handler, qt will still
quit regardless of what we do in that handler--qt will catch any exceptions we
throw, and there is no return value from the message handler.

The reason creating a QApplication will typically fail is if a window system
is not running, we may still be able to do something useful if we disable gui
support, so what we can do is start another br process with "-useGui 0" added
to the arugments (or otherwise we could start another process before trying to
create the QApplication, to see if QApplication is creatable).