ProcessWrapper round 2
In a shocking turn of events, Qt is annoying, and not particularly
consistent across platforms.

Lot of not particularly intelligent business with signals to get things to
execute on the main thread.
I would rather not even use the event driven interfaces to
QLocalServer/QLocalSocket but sadly it seems that on windows even if we
try and use the blocking calls, we are still required to use using exactly one
thread per object. Of course that makes sense since obviously we are writing
a simple event-driven GUI application.