In enroll, apply Stream's staging logic to the base algorithm
When setting up the stream in enroll, use the Stream interface's staging logic
on the input algorithm. Making this happen is slightly complex when we don't
particularly have header files, but on the plus side algorithms incorporating
something like Show (e.g. ShowFaceDetection) will not be used as a single
threaded stage in the process.