Add support for a simple append flag for enroll
This change consolidates the previous 'read' and 'noDuplicates' flags into a
single 'append' flag. If append is specified, and an output gallery already
exists, and the gallery format supports read/write or has explicit append
support, then enrollment will be restricted to those files in the input list
not already present in the gallery, and the results will be appended to the
existing gallery.

append defaults to 'false', which is a deparature from previous behavior.

The .gal format has explicit append support, for other cases, if the gallery
supports both read and write (less common than you might think), we support
append by reading the existing gallery, and writing back out to an overwriting
file. It should be possible to add explicit append support to several other
gallery types.