Fix a periodic crash when running br -objects without -useGui 0
Fix a crash periodically manifested in br -objects which was caused by
ShowTrainingTransform's transform * creating and deleting a ShowTransform so
quickly that its associated window still had incoming messages when it was
deleted in ShowTransform's destructor.

By using the deleteLater slot instead of calling delete directly we can be sure
that incoming events are processed before the window is deleted.