Add mutexes to some functions that return pointers to static variables
br_about, br_version, and br_sdk_path return pointers to static memory declared
at function scope, that is intialized once and never changed.

In versions of visual studio (prior to 2013 apparently), this initialization is
not thread safe (thread safety of this operation is in c++-11).

Removed references to managed_return_value in the documentation of these
methods. The memory returned here is not invalidated by subsequent calls to the
function, so that comment does not apply.