Add explicit multi-process support to enroll and altCompare, some bug fixes
Add explicit multi-process support to enroll and compare, if -multiProcess true
is specified, ProcessWrapper will be attached to the current algorithm in
enrollment and comparison (causing the job to be distributed over multiple
br processes instead of using multi-threading).

This is desirable since putting ProcessWrapper directly on algorithms leads to
some compositionality problems (it's very undesirable to have nested process
wrappers).

One drawback is that in multi-process mode, altcompare has to write the
explicitly enrolled gallery to a file (instead of using a memGallery) since the
slave processes obviously don't have a shared memory space. A better approach
might be to have processWrapper serialize and transmit its transform (instead
of constructing it), but even that won't work directly since transforms often
rely on reconstructing data from other files even if serialized.