Take the ground truth metadata key as input to evalClustering, default to Label
Defaulting to ClusterID did not make much sense, since the key being indexed is
in fact the ground truth, not the cluster labels (cluster labels are stored as
indices in a csv file, not a gallery).

Also, use Format to read similarity matrices instead of assuming mtx is in use.