Remove FileList return from br::Enroll
With this, we will not have to accumulate the file metadata during enroll
which will help with memory use on large jobs.

Functions actually using the return type (gui/classifier.cpp, and gallerytoolbar.cpp)
updated to avoid using it directly. In gallerytoolbar.cpp, drop support for
implicit enrollment to a .gal file.